+++ src/main/java/pl/thedeem/intellij/common/Icons.java
@@ -1,0 +1,9 @@
+package pl.thedeem.intellij.common;
+
+import com.intellij.openapi.util.IconLoader;
+
+import javax.swing.*;
+
+public interface Icons {
+    Icon DYNATRACE_LOGO = IconLoader.getIcon("/icons/dynatrace.png", Icons.class);
+}