+++ src/main/java/pl/thedeem/intellij/common/IntelliJUtils.java
@@ -20,6 +20,7 @@ import com.intellij.openapi.fileTypes.PlainTextLanguage;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectUtil;
 import com.intellij.openapi.ui.Messages;
+import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.openapi.vfs.VfsUtilCore;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -86,6 +87,14 @@ public class IntelliJUtils {
             return null;
         }
         return VfsUtilCore.getRelativePath(file, baseDir, '/');
+    }
+
+    public static @Nullable VirtualFile getProjectRelativeFile(@NotNull String relativePath, @NotNull Project project) {
+        VirtualFile baseDir = ProjectUtil.guessProjectDir(project);
+        if (baseDir == null) {
+            return null;
+        }
+        return baseDir.findFileByRelativePath(FileUtil.toSystemIndependentName(relativePath));
     }
 
     public static class StandardEditorCustomization implements EditorCustomization {