+++ src/main/java/pl/thedeem/intellij/common/StandardItemPresentation.java
@@ -2,6 +2,7 @@ package pl.thedeem.intellij.common;
 
 import com.intellij.navigation.ItemPresentation;
 import com.intellij.psi.PsiElement;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
@@ -11,11 +12,11 @@ public class StandardItemPresentation implements ItemPresentation {
     private final PsiElement element;
     private final Icon icon;
 
-    public StandardItemPresentation(String name, Icon icon) {
+    public StandardItemPresentation(@NotNull String name, @Nullable Icon icon) {
         this(name, null, icon);
     }
 
-    public StandardItemPresentation(String name, PsiElement element, Icon icon) {
+    public StandardItemPresentation(@NotNull String name, @Nullable PsiElement element, @Nullable Icon icon) {
         this.name = name;
         this.element = element;
         this.icon = icon;