+++ src/main/java/pl/thedeem/intellij/common/components/SimpleDataProviderPanel.java
@@ -1,0 +1,7 @@
+package pl.thedeem.intellij.common.components;
+
+import com.intellij.openapi.actionSystem.UiDataProvider;
+import com.intellij.util.ui.components.BorderLayoutPanel;
+
+public abstract class SimpleDataProviderPanel extends BorderLayoutPanel implements UiDataProvider {
+}