+++ src/main/java/pl/thedeem/intellij/common/services/ManagedServiceGroup.java
@@ -1,0 +1,7 @@
+package pl.thedeem.intellij.common.services;
+
+import com.intellij.execution.services.ServiceViewDescriptor;
+import com.intellij.openapi.Disposable;
+
+public interface ManagedServiceGroup extends ServiceViewDescriptor, Disposable {
+}