+++ src/main/java/pl/thedeem/intellij/dql/exec/DQLExecutionService.java
@@ -5,6 +5,7 @@ import com.intellij.ide.ActivityTracker;
 import com.intellij.ide.passwordSafe.PasswordSafe;
 import com.intellij.navigation.ItemPresentation;
 import com.intellij.openapi.actionSystem.*;
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.fileEditor.FileEditorManager;
@@ -15,33 +16,41 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jsoup.internal.StringUtil;
 import pl.thedeem.intellij.common.StandardItemPresentation;
+import pl.thedeem.intellij.common.components.InformationComponent;
 import pl.thedeem.intellij.common.sdk.DynatraceRestClient;
 import pl.thedeem.intellij.common.sdk.model.DQLExecutePayload;
 import pl.thedeem.intellij.common.sdk.model.DQLExecuteResponse;
 import pl.thedeem.intellij.common.sdk.model.DQLPollResponse;
+import pl.thedeem.intellij.common.services.ManagedService;
+import pl.thedeem.intellij.common.services.ManagedServiceGroup;
+import pl.thedeem.intellij.common.services.ProjectServicesManager;
 import pl.thedeem.intellij.dql.DQLBundle;
 import pl.thedeem.intellij.dql.DQLIcon;
 import pl.thedeem.intellij.dql.DQLUtil;
 import pl.thedeem.intellij.dql.definition.model.QueryConfiguration;
-import pl.thedeem.intellij.dql.editor.actions.ExecutionManagerAction;
+import pl.thedeem.intellij.dql.editor.actions.QueryConfigurationAction;
 import pl.thedeem.intellij.dql.exec.panel.DQLExecutionErrorPanel;
 import pl.thedeem.intellij.dql.exec.panel.DQLExecutionResult;
 import pl.thedeem.intellij.dql.fileProviders.DQLResultVirtualFile;
+import pl.thedeem.intellij.dql.services.query.DQLQueryConfigurationService;
 import pl.thedeem.intellij.dql.services.query.DQLQueryParserService;
-import pl.thedeem.intellij.dql.services.ui.DQLManagedService;
-import pl.thedeem.intellij.dql.services.ui.DQLServicesManager;
+import pl.thedeem.intellij.dql.services.ui.TenantServiceGroup;
 import pl.thedeem.intellij.dql.settings.tenants.DynatraceTenant;
 import pl.thedeem.intellij.dql.settings.tenants.DynatraceTenantsService;
 
 import javax.swing.*;
 import java.time.Instant;
 import java.time.ZoneId;
+import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
-public class DQLExecutionService implements DQLManagedService<QueryConfiguration>, UiDataProvider {
+public class DQLExecutionService implements ManagedService, UiDataProvider {
+    public static final DataKey<DQLExecutionService> EXECUTION_SERVICE = DataKey.create("executionService");
+
     private static final Logger logger = Logger.getInstance(DQLExecutionService.class);
     private final static int POLL_INTERVAL = 200;
 
@@ -52,27 +61,42 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
     private final AtomicReference<Boolean> stopping = new AtomicReference<>(false);
     private final AtomicReference<Boolean> loading = new AtomicReference<>(false);
     private final DQLExecutionResult resultPanel;
+    private final QueryConfiguration configuration;
+    private final QueryConfiguration configurationCopy;
 
     private String requestToken;
     private DQLPollResponse result;
     private String executionId;
-    private QueryConfiguration configuration;
     private DefaultActionGroup actions;
 
-    public DQLExecutionService(@NotNull String name, @NotNull Project project, @NotNull DQLProcessHandler processHandler) {
+    public DQLExecutionService(@NotNull String name, @NotNull QueryConfiguration params, @NotNull Project project, @NotNull DQLProcessHandler processHandler) {
         this.project = project;
         this.name = name;
         this.processHandler = processHandler;
+        this.configuration = params;
+        this.configurationCopy = params.copy();
         this.resultPanel = new DQLExecutionResult(project);
-        this.configuration = new QueryConfiguration();
+        this.resultPanel.setQueryConfiguration(params);
     }
 
     @Override
-    public void startExecution(@NotNull QueryConfiguration params) {
-        this.resultPanel.setQueryConfiguration(params);
+    public void dispose() {
+        stopExecution();
+    }
+
+    @Override
+    public @NotNull String getServiceId() {
+        return name;
+    }
+
+    @Override
+    public @NotNull List<ManagedServiceGroup> getParentGroups() {
+        return List.of(new TenantServiceGroup(configuration.tenant()));
+    }
+
+    public void startExecution() {
         this.resultPanel.setExecutionTime(Instant.now().atZone(ZoneId.systemDefault()));
-        this.configuration = params;
-        if (params.query() == null) {
+        if (configuration.query() == null) {
             resultPanel.show(new DQLExecutionErrorPanel(DQLBundle.message("services.executeDQL.errors.invalidPayload")));
             return;
         }
@@ -83,36 +107,41 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
         }
         String apiToken = PasswordSafe.getInstance().getPassword(DQLUtil.createCredentialAttributes(tenant.getCredentialId()));
         DynatraceRestClient client = new DynatraceRestClient(tenant.getUrl());
-        DQLExecutePayload payload = this.preparePayload(params, project);
+        DQLExecutePayload payload = this.preparePayload(configuration, project);
         this.executionId = String.valueOf(payload.getQuery().hashCode());
         logger.info(String.format("Executing a DQL query with hash %s on the tenant: %s", payload.getQuery().hashCode(), tenant.getName()));
-        requestToken = this.executeApiCall(() -> {
-            this.loading.set(true);
-            DQLExecuteResponse executedResponse = client.executeDQL(payload, apiToken);
-            logger.info(String.format(
-                    "The DQL query %s for the tenant %s was started to be executed and has a request token: %s (state: %s)",
-                    this.executionId,
-                    tenant.getName(),
-                    executedResponse.getRequestToken(),
-                    executedResponse.getState()
-            ));
-            String token = executedResponse.getRequestToken();
-            resultPanel.update(executedResponse);
-            return token;
-        });
+        this.loading.set(true);
+        ApplicationManager.getApplication().executeOnPooledThread(() -> {
+            requestToken = this.executeApiCall(() -> {
+                DQLExecuteResponse executedResponse = client.executeDQL(payload, apiToken);
+                logger.info(String.format(
+                        "The DQL query %s for the tenant %s was started to be executed and has a request token: %s (state: %s)",
+                        this.executionId,
+                        tenant.getName(),
+                        executedResponse.getRequestToken(),
+                        executedResponse.getState()
+                ));
+                String token = executedResponse.getRequestToken();
+                resultPanel.update(executedResponse);
+                return token;
+            });
 
-        if (requestToken != null) {
-            pollingFutureRef.set(startPollingRequest(requestToken, client, apiToken));
-        }
+            if (requestToken != null) {
+                pollingFutureRef.set(startPollingRequest(requestToken, client, apiToken));
+            }
+        });
     }
 
-    @Override
     public void stopExecution() {
         DynatraceTenant tenant = DynatraceTenantsService.getInstance().findTenant(configuration != null ? configuration.tenant() : null);
         if (tenant == null || executionId == null) {
             return;
         }
         stopping.set(true);
+        resultPanel.show(new InformationComponent(
+                DQLBundle.message("services.executeDQL.cancelRequested", this.requestToken),
+                AllIcons.General.Information)
+        );
         ScheduledFuture<?> scheduledFuture = pollingFutureRef.get();
         if (scheduledFuture != null && !scheduledFuture.isCancelled() && !scheduledFuture.isDone() && this.requestToken != null) {
             scheduledFuture.cancel(true);
@@ -141,7 +170,6 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
         stopping.set(false);
     }
 
-    @Override
     public boolean isRunning() {
         if (stopping.get() || loading.get()) {
             return true;
@@ -160,7 +188,7 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
         if (actions == null) {
             actions = new DefaultActionGroup();
             actions.setInjectedContext(true);
-            actions.addAction(new ExecutionManagerAction(this));
+            actions.addAction(new QueryConfigurationAction());
             actions.addSeparator();
             actions.addAction(resultPanel.getToolbarActions());
             actions.addSeparator();
@@ -173,7 +201,7 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
                     }
                     FileEditorManager.getInstance(project)
                             .openFile(new DQLResultVirtualFile(
-                                    DQLBundle.message("components.queryDetails.fileName", getName()),
+                                    DQLBundle.message("components.queryDetails.fileName", getPresentation().getPresentableText()),
                                     result
                             ), true);
                 }
@@ -191,7 +219,8 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
             actions.addAction(new AnAction(DQLBundle.message("action.DQL.CloseDQLResult.text"), null, AllIcons.General.Close) {
                 @Override
                 public void actionPerformed(@NotNull AnActionEvent anActionEvent) {
-                    DQLServicesManager.getInstance(project).stopExecution(DQLExecutionService.this);
+                    stopExecution();
+                    ProjectServicesManager.getInstance(project).unregisterService(DQLExecutionService.this);
                 }
 
                 @Override
@@ -209,18 +238,12 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
     }
 
     @Override
-    public @NotNull String getName() {
-        return name;
-    }
-
-    @Override
-    public @Nullable String getExecutionId() {
-        return this.executionId;
-    }
-
-    @Override
     public void uiDataSnapshot(@NotNull DataSink dataSink) {
         dataSink.set(EXECUTION_SERVICE, this);
+        dataSink.set(DQLQueryConfigurationService.DATA_QUERY_CONFIGURATION, configurationCopy);
+        dataSink.set(QueryConfigurationAction.SHOW_TIMEFRAME, false);
+        dataSink.set(QueryConfigurationAction.SHOW_TENANT_SELECTION, false);
+        dataSink.set(QueryConfigurationAction.SHOW_CONFIGURATION, false);
     }
 
     public @Nullable DQLPollResponse getResult() {
@@ -307,5 +330,21 @@ public class DQLExecutionService implements DQLManagedService<QueryConfiguration
         result.setMaxResultRecords(configuration.maxResultRecords());
         result.setTimezone(ZoneId.systemDefault().getId());
         return result;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) return false;
+        DQLExecutionService that = (DQLExecutionService) o;
+        return Objects.equals(configuration.tenant(), that.configuration.tenant())
+                && Objects.equals(getServiceId(), that.getServiceId());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(
+                configuration.tenant(),
+                getServiceId()
+        );
     }
 }